+++ src/fields/OptimizedImages.php
@@ -25,6 +25,7 @@ use nystudio107\imageoptimize\fields\OptimizedImages as OptimizedImagesField;
 use nystudio107\imageoptimize\gql\types\generators\OptimizedImagesGenerator;
 use nystudio107\imageoptimize\ImageOptimize;
 use nystudio107\imageoptimize\models\OptimizedImage;
+use nystudio107\imageoptimize\models\Settings;
 use ReflectionClass;
 use ReflectionException;
 use Twig\Error\LoaderError;
@@ -140,6 +141,7 @@ class OptimizedImages extends Field
 
         // Handle cases where the plugin has been uninstalled
         if (ImageOptimize::$plugin !== null) {
+            /** @var ?Settings $settings */
             $settings = ImageOptimize::$plugin->getSettings();
             if ($settings) {
                 if (empty($this->variants)) {
@@ -178,7 +180,7 @@ class OptimizedImages extends Field
                     'ignoreFilesOfType',
                     'variants',
                 ],
-                ArrayValidator::class
+                ArrayValidator::class,
             ],
         ]);
 
@@ -233,7 +235,7 @@ class OptimizedImages extends Field
                  * via a queue job to prevent it from blocking
                  */
                 ImageOptimize::$plugin->optimizedImages->resaveAsset($element->id);
-            } else if (!$request->isConsoleRequest && $request->getPathInfo() === 'actions/assets/save-image') {
+            } elseif (!$request->isConsoleRequest && $request->getPathInfo() === 'actions/assets/save-image') {
                 /**
                  * If it's not a newly uploaded/created Asset, check to see if the image
                  * itself is being updated (via the ImageEditor). If so, update the
@@ -267,7 +269,7 @@ class OptimizedImages extends Field
             $model = $value;
         } else {
             // Just create a new empty model
-            $model = new OptimizedImage(null);
+            $model = new OptimizedImage([]);
         }
 
         return $model;
@@ -511,7 +513,7 @@ class OptimizedImages extends Field
     protected function volumeHasField(Volume $volume, string $fieldHandle): bool
     {
         $result = false;
-        /** @var FieldLayout $fieldLayout */
+        /** @var ?FieldLayout $fieldLayout */
         $fieldLayout = $volume->getFieldLayout();
         // Loop through the fields in the layout to see if there is an OptimizedImages field
         if ($fieldLayout) {