+++ src/models/OptimizedImage.php
@@ -10,16 +10,15 @@
 
 namespace nystudio107\imageoptimize\models;
 
-use Composer\Installer\PackageEvent;
-use nystudio107\imageoptimize\ImageOptimize;
-use nystudio107\imageoptimize\helpers\UrlHelper;
-use nystudio107\imageoptimize\helpers\Color as ColorHelper;
-
 use Craft;
 use craft\base\Model;
-use craft\helpers\Html;
 use craft\helpers\Template;
 use craft\validators\ArrayValidator;
+use nystudio107\imageoptimize\helpers\Color as ColorHelper;
+use nystudio107\imageoptimize\helpers\UrlHelper;
+use nystudio107\imageoptimize\ImageOptimize;
+use Twig\Markup;
+use function strlen;
 
 /**
  * @author    nystudio107
@@ -87,12 +86,12 @@ class OptimizedImage extends Model
     public $lightness;
 
     /**
-     * @var int The width of the placeholder image
+     * @var ?int The width of the placeholder image
      */
     public $placeholderWidth;
 
     /**
-     * @var int The height of the placeholder image
+     * @var ?int The height of the placeholder image
      */
     public $placeholderHeight;
 
@@ -132,9 +131,9 @@ class OptimizedImage extends Model
      *
      * @param int $width
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function src(int $width = 0): string
+    public function src(int $width = 0): Markup
     {
         if (empty($width)) {
             return Template::raw(reset($this->optimizedImageUrls));
@@ -148,7 +147,7 @@ class OptimizedImage extends Model
      *
      * @param int $width
      *
-     * @return null|string|\Twig\Markup
+     * @return string
      */
     public function getSrc(int $width = 0): string
     {
@@ -161,9 +160,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcset(bool $dpr = false): string
+    public function srcset(bool $dpr = false): Markup
     {
         return Template::raw($this->getSrcsetFromArray($this->optimizedImageUrls, $dpr));
     }
@@ -187,9 +186,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetWidth(int $width, bool $dpr = false): string
+    public function srcsetWidth(int $width, bool $dpr = false): Markup
     {
         $subset = $this->getSrcsetSubsetArray($this->optimizedImageUrls, $width, 'width');
 
@@ -204,9 +203,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetMinWidth(int $width, bool $dpr = false): string
+    public function srcsetMinWidth(int $width, bool $dpr = false): Markup
     {
         $subset = $this->getSrcsetSubsetArray($this->optimizedImageUrls, $width, 'minwidth');
 
@@ -220,9 +219,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetMaxWidth(int $width, bool $dpr = false): string
+    public function srcsetMaxWidth(int $width, bool $dpr = false): Markup
     {
         $subset = $this->getSrcsetSubsetArray($this->optimizedImageUrls, $width, 'maxwidth');
 
@@ -235,9 +234,9 @@ class OptimizedImage extends Model
      *
      * @param int $width
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcWebp(int $width = 0): string
+    public function srcWebp(int $width = 0): Markup
     {
         if (empty($width)) {
             return Template::raw(reset($this->optimizedWebPImageUrls));
@@ -264,9 +263,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetWebp(bool $dpr = false): string
+    public function srcsetWebp(bool $dpr = false): Markup
     {
         return Template::raw($this->getSrcsetFromArray($this->optimizedWebPImageUrls, $dpr));
     }
@@ -290,9 +289,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetWidthWebp(int $width, bool $dpr = false): string
+    public function srcsetWidthWebp(int $width, bool $dpr = false): Markup
     {
         $subset = $this->getSrcsetSubsetArray($this->optimizedWebPImageUrls, $width, 'width');
 
@@ -307,9 +306,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetMinWidthWebp(int $width, bool $dpr = false): string
+    public function srcsetMinWidthWebp(int $width, bool $dpr = false): Markup
     {
         $subset = $this->getSrcsetSubsetArray($this->optimizedWebPImageUrls, $width, 'minwidth');
 
@@ -324,9 +323,9 @@ class OptimizedImage extends Model
      * @param bool $dpr Whether to generate 1x, 2x srcsets vs the normal XXXw
      *                  srcsets
      *
-     * @return \Twig\Markup|null
+     * @return Markup
      */
-    public function srcsetMaxWidthWebp(int $width, bool $dpr = false): string
+    public function srcsetMaxWidthWebp(int $width, bool $dpr = false): Markup
     {
         $subset = $this->getSrcsetSubsetArray($this->optimizedWebPImageUrls, $width, 'maxwidth');
 
@@ -384,159 +383,37 @@ class OptimizedImage extends Model
      * Generate a complete <link rel="preload"> tag for this OptimizedImages model
      * ref: https://web.dev/preload-responsive-images/#imagesrcset-and-imagesizes
      *
-     * @param array $linkAttrs
-     *
-     * @return \Twig\Markup
+     * @return LinkPreloadTag
      */
-    public function linkPreloadTag($linkAttrs = [])
+    public function linkPreloadTag(): LinkPreloadTag
     {
-        // Any web browser that supports link rel="preload" as="image" also supports webp, so prefer that
-        $srcset = $this->optimizedImageUrls;
-        if (!empty($this->optimizedWebPImageUrls)) {
-            $srcset = $this->optimizedWebPImageUrls;
-        }
-        // Merge the passed in options with the tag attributes
-        $attrs = array_merge([
-            'rel' => 'preload',
-            'as' => 'image',
-            'href' => reset($srcset),
-            'imagesrcset' => $this->getSrcsetFromArray($srcset),
-            'imagesizes' => '100vw',
-        ],
-            $linkAttrs
-        );
-        // Remove any empty attributes
-        $attrs = array_filter($attrs);
-        // Render the tag
-        $tag = Html::tag('link', '', $attrs);
-
-        return Template::raw($tag);
+        return new LinkPreloadTag(['optimizedImage' => $this]);
     }
 
     /**
-     * Generate a complete <img> tag for this OptimizedImages model
+     * Generate a complete <img> tag for this OptimizedImage model
      *
-     * @param string $loading 'eager', 'lazy', 'lazySizes', 'lazySizesFallback'
-     * @param string $placeHolder 'box', 'color', 'image', 'silhouette'
-     * @param array $imgAttrs
-     *
-     * @return \Twig\Markup
+     * @return ImgTag
      */
-    public function imgTag($loading = 'eager', $placeHolder = 'box', $imgAttrs = [])
+    public function imgTag(): ImgTag
     {
-        // Merge the passed in options with the tag attributes
-        $attrs = array_merge([
-            'class' => '',
-            'style' => '',
-            'width' => $this->placeholderWidth,
-            'height' => $this->placeholderHeight,
-            'src' => reset($this->optimizedImageUrls),
-            'srcset' => $this->getSrcsetFromArray($this->optimizedImageUrls),
-            'sizes' => '100vw',
-            'loading' => '',
-        ],
-            $imgAttrs
-        );
-        // Handle lazy loading
-        if ($loading !== 'eager') {
-            $attrs = $this->swapLazyLoadAttrs($loading, $placeHolder, $attrs);
-        }
-        // Remove any empty attributes
-        $attrs = array_filter($attrs);
-        // Render the tag
-        $tag = Html::tag('img', '', $attrs);
-
-        return Template::raw($tag);
+        return new ImgTag(['optimizedImage' => $this]);
     }
 
     /**
-     * Generate a complete <picture> tag for this OptimizedImages model
+     * Generate a complete <picture> tag for this OptimizedImage model
      *
-     * @param string $loading 'eager', 'lazy', 'lazySizes', 'lazySizesFallback'
-     * @param string $placeHolder 'box', 'color', 'image', 'silhouette'
-     * @param array $pictureAttrs
-     * @param array $srcsetAttrs
-     * @param array $imgAttrs
-     *
-     * @return \Twig\Markup
+     * @return PictureTag
      */
-    public function pictureTag($loading = 'eager', $placeHolder = 'box', $pictureAttrs = [], $srcsetAttrs = [], $imgAttrs = [])
+    public function pictureTag(): PictureTag
     {
-        $content = '';
-        // Handle the webp srcset
-        if (!empty($this->optimizedWebPImageUrls)) {
-            // Merge the passed in options with the tag attributes
-            $attrs = array_merge([
-                'srcset' => $this->getSrcsetFromArray($this->optimizedWebPImageUrls),
-                'type' => 'image/webp',
-                'sizes' => '100vw',
-            ],
-                $srcsetAttrs
-            );
-            // Handle lazy loading
-            if ($loading !== 'eager') {
-                $attrs = $this->swapLazyLoadAttrs($loading, $placeHolder, $attrs);
-            }
-            // Remove any empty attributes
-            $attrs = array_filter($attrs);
-            // Render the tag
-            $content .= Html::tag('source', '', $attrs);
-        }
-        // Handle the regular srcset
-        if (!empty($this->optimizedImageUrls)) {
-            // Merge the passed in options with the tag attributes
-            $attrs = array_merge([
-                'srcset' => $this->getSrcsetFromArray($this->optimizedImageUrls),
-                'sizes' => '100vw',
-            ],
-                $srcsetAttrs
-            );
-            // Handle lazy loading
-            if ($loading !== 'eager') {
-                $attrs = $this->swapLazyLoadAttrs($loading, $placeHolder, $attrs);
-            }
-            // Remove any empty attributes
-            $attrs = array_filter($attrs);
-            // Render the tag
-            $content .= Html::tag('source', '', $attrs);
-        }
-        // Handle the img tag
-        /** @noinspection SuspiciousAssignmentsInspection */
-        $attrs = array_merge([
-            'class' => '',
-            'style' => '',
-            'width' => $this->placeholderWidth,
-            'height' => $this->placeholderHeight,
-            'src' => reset($this->optimizedImageUrls),
-            'loading' => '',
-        ],
-            $imgAttrs
-        );
-        // Handle lazy loading
-        if ($loading !== 'eager') {
-            $attrs = $this->swapLazyLoadAttrs($loading, $placeHolder, $attrs);
-        }
-        // Remove any empty attributes
-        $attrs = array_filter($attrs);
-        // Render the tag
-        $content .= Html::tag('img', '', $attrs);
-        // Merge the passed in options with the tag attributes
-        $attrs = array_merge([
-        ],
-            $pictureAttrs
-        );
-        // Remove any empty attributes
-        $attrs = array_filter($attrs);
-        // Render the tag
-        $tag = Html::tag('picture', $content, $attrs);
-
-        return Template::raw($tag);
+        return new PictureTag(['optimizedImage' => $this]);
     }
 
     /**
      * Return a base64-encoded placeholder image
      *
-     * @return \Twig\Markup|null
+     * @return Markup|null
      */
     public function placeholderImage()
     {
@@ -567,7 +444,7 @@ class OptimizedImage extends Model
     public function placeholderImageSize(): string
     {
         $placeholder = $this->placeholderImage();
-        $contentLength = !empty(\strlen($placeholder)) ? \strlen($placeholder) : 0;
+        $contentLength = !empty(strlen($placeholder)) ? strlen($placeholder) : 0;
 
         return ImageOptimize::$plugin->optimize->humanFileSize($contentLength, 1);
     }
@@ -577,7 +454,7 @@ class OptimizedImage extends Model
      *
      * @param string|null $color
      *
-     * @return \Twig\Markup|null
+     * @return Markup|null
      */
     public function placeholderBox(string $color = null)
     {
@@ -606,7 +483,7 @@ class OptimizedImage extends Model
     public function placeholderBoxSize(): string
     {
         $placeholder = $this->placeholderBox();
-        $contentLength = !empty(\strlen($placeholder)) ? \strlen($placeholder) : 0;
+        $contentLength = !empty(strlen($placeholder)) ? strlen($placeholder) : 0;
 
         return ImageOptimize::$plugin->optimize->humanFileSize($contentLength, 1);
     }
@@ -614,7 +491,7 @@ class OptimizedImage extends Model
     /**
      * Return a silhouette of the image as an SVG placeholder
      *
-     * @return \Twig\Markup|null
+     * @return Markup|null
      */
     public function placeholderSilhouette()
     {
@@ -645,7 +522,7 @@ class OptimizedImage extends Model
     public function placeholderSilhouetteSize(): string
     {
         $placeholder = $this->placeholderSilhouette();
-        $contentLength = !empty(\strlen($placeholder)) ? \strlen($placeholder) : 0;
+        $contentLength = !empty(strlen($placeholder)) ? strlen($placeholder) : 0;
 
         return ImageOptimize::$plugin->optimize->humanFileSize($contentLength, 1);
     }
@@ -700,6 +577,34 @@ class OptimizedImage extends Model
         return ImageOptimize::$plugin->optimize->humanFileSize($contentLength, 1);
     }
 
+    /**
+     * @param array $array
+     * @param bool $dpr
+     *
+     * @return string
+     */
+    public function getSrcsetFromArray(array $array, bool $dpr = false): string
+    {
+        $srcset = '';
+        foreach ($array as $key => $value) {
+            if ($dpr) {
+                $descriptor = '1x';
+                if (!empty($array[(int)$key / 2])) {
+                    $descriptor = '2x';
+                }
+                if (!empty($array[(int)$key / 3])) {
+                    $descriptor = '3x';
+                }
+            } else {
+                $descriptor = $key . 'w';
+            }
+            $srcset .= $value . ' ' . $descriptor . ', ';
+        }
+        $srcset = rtrim($srcset, ', ');
+
+        return $srcset;
+    }
+
     // Protected Methods
     // =========================================================================
 
@@ -741,39 +646,11 @@ class OptimizedImage extends Model
     }
 
     /**
-     * @param array $array
-     * @param bool $dpr
-     *
-     * @return string
-     */
-    protected function getSrcsetFromArray(array $array, bool $dpr = false): string
-    {
-        $srcset = '';
-        foreach ($array as $key => $value) {
-            if ($dpr) {
-                $descriptor = '1x';
-                if (!empty($array[(int)$key / 2])) {
-                    $descriptor = '2x';
-                }
-                if (!empty($array[(int)$key / 3])) {
-                    $descriptor = '3x';
-                }
-            } else {
-                $descriptor = $key . 'w';
-            }
-            $srcset .= $value . ' ' . $descriptor . ', ';
-        }
-        $srcset = rtrim($srcset, ', ');
-
-        return $srcset;
-    }
-
-    /**
      * Return a default placeholder image
      *
-     * @return \Twig\Markup
+     * @return Markup
      */
-    protected function defaultPlaceholderImage(): \Twig\Markup
+    protected function defaultPlaceholderImage(): Markup
     {
         $width = 1;
         $height = 1;